Move to separate files, and add SimpleDecodable
